+++ src/LOADFDATA/make_munuDipY.f90
@@ -1,8 +1,9 @@
 subroutine make_munuDipY ()
+  use iso_c_binding
   use M_fdata, only: ME2cDipY_max, nspecies, nssh, num_orb, lssh, &
     & index_max2cDipY, ME2c_max, muDipY, nuDipY
   implicit none
-  integer :: index, in1, in2, issh1, issh2, l1, l2, n1, n2
+  integer(c_long) :: index, in1, in2, issh1, issh2, l1, l2, n1, n2
 
   ME2cDipY_max = 0
   allocate (index_max2cDipY (nspecies, nspecies))
