+++ src/SYSTEM/getenergy.f90
@@ -1,8 +1,9 @@
 subroutine getenergy () 
+  use iso_c_binding
   use M_system 
   use M_fdata, only : nssh
   implicit none
-  integer :: iatom, issh
+  integer(c_long) :: iatom, issh
 
   call get_ewald (iforce)
